front end modified in the widget
